+++ packages/talker/test/talker_stream_test.dart
@@ -1,0 +1,14 @@
+import 'package:talker/talker.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('TalkerStream', () {
+    final talker = Talker();
+
+    test('handle', () async {
+      talker.error('Test message');
+      talker.stream
+          .listen((log) => expectAsync1((event) => event is TalkerLog));
+    });
+  });
+}