+++ packages/talker_flutter/lib/src/ui/talker_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:group_button/group_button.dart';
 import 'package:talker_flutter/src/controller/talker_screen_controller.dart';
-import 'package:talker_flutter/src/widgets/widgets.dart';
+import 'package:talker_flutter/src/ui/ui.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 /// UI view for output of all Talker logs and errors