+++ utils/publish_talker.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+cd packages/talker
+flutter pub publish