+++ packages/talker_logger/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.11.1
+- **FEAT**: Move back copyWith method
+
+## 0.11.0
+- **INFO**: Make 100% tests coverage
+- **INFO**: Update package docs, examples, tests 
+
 ## 0.9.0
 - **INFO**: Make small refactor